requirements.txt
Django==3.2.15
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.14.0
psycopg2-binary==2.9.3
python-decouple==3.6